feat: created context for habitat features
- created hook for using the context
- additional styled components and containers
