+++ app/src/features/surveys/SurveyRouter.tsx
@@ -12,6 +12,7 @@ import React from 'react';
 import { Redirect, Switch } from 'react-router';
 import RouteWithTitle from 'utils/RouteWithTitle';
 import { getTitle } from 'utils/Utils';
+import { HabitatFeatureRouter } from './habitat-features/SurveyHabitatFeatureRouter';
 import { TelemetryRouter } from './telemetry/TelemetryRouter';
 
 /**
@@ -70,6 +71,21 @@ const SurveyRouter: React.FC = () => {
           ]}
           validSystemRoles={[SYSTEM_ROLE.SYSTEM_ADMIN, SYSTEM_ROLE.DATA_ADMINISTRATOR]}>
           <TelemetryRouter />
+        </ProjectRoleRouteGuard>
+      </RouteWithTitle>
+
+      {/* Habitat Features Routes */}
+      <RouteWithTitle
+        path="/admin/projects/:id/surveys/:survey_id/habitat-features"
+        title={getTitle('Manage Habitat Features')}>
+        <ProjectRoleRouteGuard
+          validProjectPermissions={[
+            PROJECT_PERMISSION.COORDINATOR,
+            PROJECT_PERMISSION.COLLABORATOR,
+            PROJECT_PERMISSION.OBSERVER
+          ]}
+          validSystemRoles={[SYSTEM_ROLE.SYSTEM_ADMIN, SYSTEM_ROLE.DATA_ADMINISTRATOR]}>
+          <HabitatFeatureRouter />
         </ProjectRoleRouteGuard>
       </RouteWithTitle>
