+++ app/src/features/surveys/habitat-features/SurveyHabitatFeaturePage.tsx
@@ -1,9 +1,17 @@
-import { CircularProgress, Stack } from '@mui/material';
+import { Box, CircularProgress, Stack } from '@mui/material';
+import { DialogContextProvider } from 'contexts/dialogContext';
 import { ProjectContext } from 'contexts/projectContext';
 import { SurveyContext } from 'contexts/surveyContext';
 import { useContext } from 'react';
 import { SurveyManagePageEnum, SurveyManagePageHeader } from '../components/SurveyManagePageHeader';
+import { SamplingSiteList } from '../observations/sampling-sites/SamplingSiteList';
+import { SurveyHabitatFeatureTableContainer } from './components/SurveyHabitatFeatureTableContainer';
 
+/**
+ * Returns the page for managing Habitat Features
+ *
+ * @return {*} {JSX.Element}
+ */
 export const SurveyHabitatFeaturePage = (): JSX.Element => {
   const surveyContext = useContext(SurveyContext);
   const projectContext = useContext(ProjectContext);
@@ -29,6 +37,26 @@ export const SurveyHabitatFeaturePage = (): JSX.Element => {
         survey_id={surveyContext.surveyId}
         survey_name={surveyContext.surveyDataLoader.data.surveyData.survey_details.survey_name}
       />
+      <Stack
+        direction="row"
+        gap={1}
+        sx={{
+          flex: '1 1 auto',
+          p: 1
+        }}>
+        {/* Sampling Site List */}
+        <Box flex="0 0 auto" width="400px">
+          <DialogContextProvider>
+            {/* TODO: Mac: Update isMenuDisabled with correct value */}
+            <SamplingSiteList isMenuDisabled={false} />
+          </DialogContextProvider>
+        </Box>
+
+        {/* Survey Habitat Feature Table */}
+        <Box flex="1 1 auto">
+          <SurveyHabitatFeatureTableContainer />
+        </Box>
+      </Stack>
     </Stack>
   );
 };
