+++ app/src/features/surveys/view/survey-spatial/components/habitat-feature/SurveySpatialHabitatFeature.tsx
@@ -1,19 +1,20 @@
 import Box from '@mui/material/Box';
 import { IStaticLayer } from 'components/map/components/StaticLayers';
+import { HabitatFeatureTableContextProvider } from 'contexts/habitatFeatureTableContext';
 import SurveyMap from 'features/surveys/view/SurveyMap';
-import { SurveyHabitatFeatureTable } from '../../../../habitat-features/components/SurveyHabitatFeatureTable';
+import { SurveySpatialHabitatFeatureTableContainer } from './SurveySpatialHabitatFeatureTableContainer';
 
-/**
- * Array of additional static layers to be added to the map.
- */
 interface ISurveySpatialHabitatFeatureProps {
+  /**
+   * Array of additional static layers to be added to the map.
+   *
+   * @type {IStaticLayer[]}
+   */
   staticLayers?: IStaticLayer[];
 }
 
 /**
  * Container displaying a map of Habitat Features in the Survey
- *
- * TODO: Mac: Implement the the fetching logic for habitat features
  *
  * @param {ISurveySpatialHabitatFeatureProps} props - The props for the component.
  * @returns {*} {JSX.Element}
@@ -26,12 +27,11 @@ export const SurveySpatialHabitatFeature = (props: ISurveySpatialHabitatFeatureP
         <SurveyMap staticLayers={[...(props.staticLayers ?? [])]} isLoading={false} />
       </Box>
 
-      {/*
-        Display data table with habitat feature details
-        TODO: Mac: Replace the isLoading with the actual loading state
-        */}
+      {/* Display data table with habitat feature details */}
       <Box height={{ xs: 300, md: 500 }} display="flex" flexDirection="column">
-        <SurveyHabitatFeatureTable isLoading={false} />
+        <HabitatFeatureTableContextProvider>
+          <SurveySpatialHabitatFeatureTableContainer />
+        </HabitatFeatureTableContextProvider>
       </Box>
     </>
   );
