+++ app/src/interfaces/useMarkdownApi.interface.ts
@@ -27,5 +27,6 @@ export enum MarkdownTypeNameEnum {
   SAMPLING_SITES = 'Sampling Sites',
   SAMPLING_PERIODS = 'Sampling Periods',
   SURVEY_METADATA = 'Survey Metadata',
-  OBSERVATIONS = 'Observations'
+  OBSERVATIONS = 'Observations',
+  HABITAT_FEATURES = 'Habitat Features'
 }
